refactor(admin-ui): Phase2-P1-2 A3 — TextInputTab をローカルから component 抽出 + CATEGORY_LABELS を shared 化 (機能変更なし) (GID 1214120401966111)
